# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Cygnus API",
    "description": "Backend API for the Cygnus project",
    "version": "0.1.0"
  },
...
  "paths": {
...
    "/users/me": {
      "patch": {
        "summary": "Update the authenticated user's profile",
        "tags": [
          "Users"
        ],
        "description": "Updates name, email and/or password for the authenticated account. currentPassword is required whenever email or password are being changed.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "description": "The parent or caregiver's full name"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "New email address. Requires currentPassword"
                  },
                  "password": {
                    "type": "string",
                    "minLength": 8,
                    "description": "New plain-text password, at least 8 characters. Requires currentPassword"
                  },
                  "currentPassword": {
                    "type": "string",
                    "minLength": 1,
                    "description": "Required whenever email or password change"
                  },
                  "emailNotificationsEnabled": {
                    "type": "boolean",
                    "description": "Whether the user wants to receive reminder emails (vaccines, appointments)"
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The authenticated user account",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "email": {
                      "type": "string",
                      "format": "email"
                    },
                    "name": {
                      "type": "string"
                    },
                    "emailNotificationsEnabled": {
                      "type": "boolean",
                      "description": "Whether the user wants to receive reminder emails (vaccines, appointments)"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  },
                  "required": [
                    "id",
                    "email",
                    "name",
                    "emailNotificationsEnabled",
                    "createdAt"
                  ],
                  "additionalProperties": false,
                  "description": "The authenticated user account"
                }
              }
            }
          },
          "400": {
            "description": "The request could not be completed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "enum": [
                        "error"
                      ]
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "status",
                    "message"
                  ],
                  "additionalProperties": false,
                  "description": "The request could not be completed"
                }
              }
            }
          },
          "401": {
            "description": "The request could not be completed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "enum": [
                        "error"
                      ]
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "status",
                    "message"
                  ],
                  "additionalProperties": false,
                  "description": "The request could not be completed"
                }
              }
            }
          },
          "409": {
            "description": "The request could not be completed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "enum": [
                        "error"
                      ]
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "status",
                    "message"
                  ],
                  "additionalProperties": false,
                  "description": "The request could not be completed"
                }
              }
            }
          },
          "500": {
            "description": "The request could not be completed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "enum": [
                        "error"
                      ]
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "status",
                    "message"
                  ],
                  "additionalProperties": false,
                  "description": "The request could not be completed"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete the authenticated user's account",
        "tags": [
          "Users"
        ],
        "description": "Permanently deletes the account after confirming currentPassword, cascading to all owned babies, their vaccine records, appointments, milestones and notifications. Clears the session cookies.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "currentPassword": {
                    "type": "string",
                    "minLength": 1,
                    "description": "The user's current password, required to confirm account deletion"
                  }
                },
                "required": [
                  "currentPassword"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Account deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "enum": [
                    "null"
                  ],
                  "nullable": true,
                  "description": "Account deleted successfully"
                }
              }
            }
          },
          "400": {
            "description": "The request could not be completed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "enum": [
                        "error"
                      ]
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "status",
                    "message"
                  ],
                  "additionalProperties": false,
                  "description": "The request could not be completed"
                }
              }
            }
          },
          "401": {
            "description": "The request could not be completed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "enum": [
                        "error"
                      ]
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "status",
                    "message"
                  ],
                  "additionalProperties": false,
                  "description": "The request could not be completed"
                }
              }
            }
          },
          "500": {
            "description": "The request could not be completed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "enum": [
                        "error"
                      ]
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "status",
                    "message"
                  ],
                  "additionalProperties": false,
                  "description": "The request could not be completed"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}